requirements.txt
flask>=3.0
psycopg2-binary>=2.9
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.40
pandas>=2.1
rapidfuzz>=3.5
python-dotenv>=1.0
lxml>=5.0
